requirements.txt
numpy<2
flask==3.1.1
torch==2.2.2
torchvision==0.17.2
Pillow==10.2.0
python-dotenv==1.0.1
transformers==4.40.0
opencv-python==4.9.0.80
gunicorn
tensorflow-cpu
ultralytics